+++ examples/demo-rollup/src/mock_rollup.rs
@@ -10,7 +10,7 @@ use sov_mock_zkvm::{MockCodeCommitment, MockZkVerifier, MockZkvm};
 use sov_modules_api::default_spec::DefaultSpec;
 use sov_modules_api::execution_mode::{ExecutionMode, Native, Zk};
 use sov_modules_api::higher_kinded_types::Generic;
-use sov_modules_api::{CryptoSpec, Spec, Zkvm};
+use sov_modules_api::{CryptoSpec, SovApiProofSerializer, Spec, Zkvm};
 use sov_modules_rollup_blueprint::pluggable_traits::PluggableSpec;
 use sov_modules_rollup_blueprint::{FullNodeBlueprint, RollupBlueprint};
 use sov_modules_stf_blueprint::{RuntimeEndpoints, StfBlueprint};
@@ -66,6 +66,8 @@ impl FullNodeBlueprint<Native> for MockDemoRollup<Native> {
             <MockDemoRollup<Zk> as RollupBlueprint<Zk>>::Kernel,
         >,
     >;
+
+    type ProofSerializer = SovApiProofSerializer<Self::Spec>;
 
     fn create_outer_code_commitment(
         &self,